Split instance mapping into diff submodules
